Add files via upload
This project focuses on customer care calls from the insurance category in ABC company. The objective is to analyze these calls over a full day to determine various metrics, such as the average number of calls, total call volume, abandon rate, and the number of additional agents required to ensure a positive customer experience.